backend/restaurant-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.1
pydantic==2.4.2
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.1
black==23.10.1
flake8==6.1.0
isort==5.12.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pymongo==4.6.1
